added xor encryption, generating prime numbers, better exception handling

+++ src/main/java/pl/pwr/common/service/encryption/caesar/operation/Operation.java
@@ -1,9 +1,11 @@
 package pl.pwr.common.service.encryption.caesar.operation;
 
+import java.math.BigInteger;
+
 /**
  * Created by Evelan on 19/10/2016.
  */
 public interface Operation {
 
-    char doOperation(char singleChar, Integer key);
+    char doOperation(char singleChar, BigInteger key);
 }
