+++ src/main/java/pl/pwr/common/service/encryption/caesar/operation/OperationContext.java
@@ -1,5 +1,7 @@
 package pl.pwr.common.service.encryption.caesar.operation;
 
+import java.math.BigInteger;
+
 /**
  * Created by Evelan on 19/10/2016.
  */
@@ -11,7 +13,7 @@ public class OperationContext {
         this.operation = operation;
     }
 
-    public char executeOperation(char singleChar, Integer key) {
+    public char executeOperation(char singleChar, BigInteger key) {
         return operation.doOperation(singleChar, key);
     }
