+++ src/main/java/pl/pwr/common/service/encryption/caesar/operation/OperationSubstract.java
@@ -1,11 +1,13 @@
 package pl.pwr.common.service.encryption.caesar.operation;
 
+import java.math.BigInteger;
+
 /**
  * Created by Evelan on 19/10/2016.
  */
 public class OperationSubstract implements Operation {
     @Override
-    public char doOperation(char singleChar, Integer key) {
-        return (char) (singleChar - key);
+    public char doOperation(char singleChar, BigInteger key) {
+        return (char) (singleChar - key.intValue());
     }
 }
